+++ client/components/LoginForm.jsx
@@ -1,0 +1,81 @@
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+const LoginForm = () => {
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+  const [isPwdWrong, setIsPwdWrong] = useState(false);
+  const [fieldsFilled, setFieldsFilled] = useState(true);
+  const [errorMsgShow, setErrorMsgShow] = useState(false);
+  const navigate = useNavigate();
+  const handleLogin = (formData) => {
+    setIsPwdWrong(false);
+    setFieldsFilled(true);
+    setErrorMsgShow(false);
+    fetch('/login', {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(formData),
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        console.log(data, 'success logged in');
+        if (data.success) {
+          navigate('/mainpage');
+        } else {
+          if (data.message === 'incorrect information') {
+            setIsPwdWrong(true);
+          } else if (data.message === 'missing information') {
+            setFieldsFilled(false);
+          } else {
+            setErrorMsgShow(true);
+          }
+        }
+      })
+      .catch((error) => {
+        setErrorMsgShow(true);
+        console.error('Error:', error);
+      });
+  };
+
+  const handleSignUpClick = () => {
+    navigate('/signup');
+  };
+
+  return (
+    <div>
+      <h2>Login</h2>
+      <form onSubmit={handleLogin}>
+        <div>
+          <label htmlFor='username'>Username:</label>
+          <input
+            type='text'
+            id='username'
+            value={username}
+            onChange={(e) => setUsername(e.target.value)}
+            required
+          />
+        </div>
+        <div>
+          <label htmlFor='password'>Password:</label>
+          <input
+            type='password'
+            id='password'
+            value={password}
+            onChange={(e) => setPassword(e.target.value)}
+            required
+          />
+        </div>
+        <button type='submit'>Login</button>
+      </form>
+      {isPwdWrong && <p>Incorrect password</p>}
+      {!fieldsFilled && <p>Please fill all fields</p>}
+      {errorMsgShow && <p>An error occurred. Please try again.</p>}
+      <button onClick={handleSignUpClick}>Sign Up</button>
+    </div>
+  );
+};
+
+export default LoginForm;
